+++ src/com/example/banban/ui/myaccount/ShareProductFragment.java
@@ -143,7 +143,7 @@ public class ShareProductFragment extends Fragment {
 		int product_id = object.getInt("product_id");
 		String product_name = object.getString("product_name");
 
-		int price = object.getInt("price");
+		double price = object.getDouble("price");
 		String image = object.getString("image");
 		String amount_spec = object.getString("amount_spec");
 		int favorites = object.getInt("favorites");
